Add \Device\KsecDD emulation and mask 32-bit syscall arguments
- Emulate \Device\KsecDD (RNG + memory encrypt/decrypt) so the bcrypt/CNG
  DllMain succeeds during Windows PE process initialization.
- Mask ULONG syscall arguments that were read as 64-bit; some ntdll builds
  leave garbage in the upper 32 bits, which broke handler argument checks.

+++ Brovan/Core/Emulation/OS/Windows/Files/NtQueryDirectoryFile.cs
@@ -20,9 +20,9 @@ namespace Brovan.Core.Emulation.OS.Windows
             ulong FileInformation = Instance.WinHelper.GetArg64(5);
             uint Length = (uint)Instance.WinHelper.GetArg64(6);
             uint FileInformationClass = (uint)Instance.WinHelper.GetArg64(7);
-            bool ReturnSingleEntry = Instance.WinHelper.GetArg64(8) != 0;
+            bool ReturnSingleEntry = (uint)Instance.WinHelper.GetArg64(8) != 0;
             ulong FileName = Instance.WinHelper.GetArg64(9);
-            bool RestartScan = Instance.WinHelper.GetArg64(10) != 0;
+            bool RestartScan = (uint)Instance.WinHelper.GetArg64(10) != 0;
 
             uint QueryFlags = 0;
             if (RestartScan)
